requirements.txt
wxpython
pillow